Fix float precision in Finetune CLI

+++ src/seamless_communication/cli/m4t/finetune/finetune.py
@@ -126,7 +126,7 @@ def main() -> None:
     args = init_parser().parse_args()
     dist_utils.init_distributed([logger, trainer.logger])
     device = torch.device("cuda")
-    float_dtype = torch.float16
+    float_dtype = torch.float32
     text_tokenizer: NllbTokenizer = load_unity_text_tokenizer(args.model_name)
     unit_tokenizer: UnitTokenizer = load_unity_unit_tokenizer(args.model_name)
     finetune_params = trainer.FinetuneParams(
